+++ guillotina/api/files.py
@@ -15,8 +15,30 @@ from guillotina.interfaces import IStaticFile
 import mimetypes
 
 
+def _traversed_file_doc(summary, parameters=[], responses={
+    "200": {
+        "description": "Successfully updated content"
+    }
+}):
+    return {
+        'traversed_service_definitions': {
+            '{field_name}': {
+                'summary': summary,
+                'parameters': [{
+                    'name': 'field_name',
+                    'in': 'path',
+                    'description': 'Name of file field',
+                    'required': True
+                }] + parameters,
+                'responses': responses
+            }
+        }
+    }
+
+
 # Static File
-@configure.service(context=IStaticFile, method='GET', permission='guillotina.AccessContent')
+@configure.service(
+    context=IStaticFile, method='GET', permission='guillotina.AccessContent')
 class FileGET(DownloadService):
     async def serve_file(self, fi):
         filepath = str(fi.file_path.absolute())
@@ -43,7 +65,8 @@ class FileGET(DownloadService):
             return await self.serve_file(self.context)
 
 
-@configure.service(context=IStaticDirectory, method='GET', permission='guillotina.AccessContent')
+@configure.service(
+    context=IStaticDirectory, method='GET', permission='guillotina.AccessContent')
 class DirectoryGET(FileGET):
     async def __call__(self):
         for possible_default_file in app_settings['default_static_filenames']:
@@ -52,8 +75,10 @@ class DirectoryGET(FileGET):
 
 
 # Field File
-@configure.service(context=IResource, method='PATCH', permission='guillotina.ModifyContent',
-                   name='@upload')
+@configure.service(
+    context=IResource, method='PATCH', permission='guillotina.ModifyContent',
+    name='@upload',
+    **_traversed_file_doc('Update the content of a file'))
 class UploadFile(TraversableFieldService):
 
     async def __call__(self):
@@ -64,8 +89,10 @@ class UploadFile(TraversableFieldService):
         return await adapter.upload()
 
 
-@configure.service(context=IResource, method='GET', permission='guillotina.ViewContent',
-                   name='@download')
+@configure.service(
+    context=IResource, method='GET', permission='guillotina.ViewContent',
+    name='@download',
+    **_traversed_file_doc('Download the content of a file'))
 class DownloadFile(TraversableDownloadService):
 
     async def __call__(self):
@@ -76,8 +103,55 @@ class DownloadFile(TraversableDownloadService):
         return await adapter.download()
 
 
-@configure.service(context=IResource, method='POST', permission='guillotina.ModifyContent',
-                   name='@tusupload')
+@configure.service(
+    context=IResource, method='POST', permission='guillotina.ModifyContent',
+    name='@tusupload',
+    **_traversed_file_doc('TUS endpoint', parameters=[{
+        "name": "Upload-Offset",
+        "in": "headers",
+        "type": "integer",
+        "required": True
+    }, {
+        "name": "UPLOAD-LENGTH",
+        "in": "headers",
+        "type": "integer",
+        "required": True
+    }, {
+        "name": "UPLOAD-MD5",
+        "in": "headers",
+        "type": "string",
+        "required": False
+    }, {
+        "name": "UPLOAD-EXTENSION",
+        "in": "headers",
+        "type": "string",
+        "required": False
+    }, {
+        "name": "TUS-RESUMABLE",
+        "in": "headers",
+        "type": "string",
+        "required": True
+    }, {
+        "name": "UPLOAD-METADATA",
+        "in": "headers",
+        "type": "string",
+        "required": False
+    }], responses={
+        '204': {
+            'description': 'Successfully patched data',
+            'headers': {
+                'Location': {
+                    'type': 'string'
+                },
+                'Tus-Resumable': {
+                    'type': 'string'
+                },
+                'Access-Control-Expose-Headers': {
+                    'type': 'string'
+                }
+            }
+        }
+    }))
 class TusCreateFile(UploadFile):
 
     async def __call__(self):
@@ -88,8 +162,25 @@ class TusCreateFile(UploadFile):
         return await adapter.tus_create()
 
 
-@configure.service(context=IResource, method='HEAD', permission='guillotina.ModifyContent',
-                   name='@tusupload')
+@configure.service(
+    context=IResource, method='HEAD', permission='guillotina.ModifyContent',
+    name='@tusupload',
+    **_traversed_file_doc('TUS endpoint', responses={
+        '200': {
+            'description': 'Successfully patched data',
+            'headers': {
+                'Upload-Offset': {
+                    'type': 'integer'
+                },
+                'Tus-Resumable': {
+                    'type': 'string'
+                },
+                'Access-Control-Expose-Headers': {
+                    'type': 'string'
+                }
+            }
+        }
+    }))
 class TusHeadFile(UploadFile):
 
     async def __call__(self):
@@ -100,8 +191,29 @@ class TusHeadFile(UploadFile):
         return await adapter.tus_head()
 
 
-@configure.service(context=IResource, method='PATCH', permission='guillotina.ModifyContent',
-                   name='@tusupload')
+@configure.service(
+    context=IResource, method='PATCH', permission='guillotina.ModifyContent',
+    name='@tusupload',
+    **_traversed_file_doc('TUS endpoint', parameters=[{
+        "name": "Upload-Offset",
+        "in": "headers",
+        "type": "integer",
+        "required": True
+    }, {
+        "name": "CONTENT-LENGTH",
+        "in": "headers",
+        "type": "integer",
+        "required": True
+    }], responses={
+        '204': {
+            'description': 'Successfully patched data',
+            'headers': {
+                'Upload-Offset': {
+                    'type': 'integer'
+                }
+            }
+        }
+    }))
 class TusPatchFile(UploadFile):
 
     async def __call__(self):
@@ -112,8 +224,28 @@ class TusPatchFile(UploadFile):
         return await adapter.tus_patch()
 
 
-@configure.service(context=IResource, method='OPTIONS', permission='guillotina.AccessPreflight',
-                   name='@tusupload')
+@configure.service(
+    context=IResource, method='OPTIONS', permission='guillotina.AccessPreflight',
+    name='@tusupload',
+    **_traversed_file_doc('TUS endpoint', responses={
+        '200': {
+            'description': 'Successfully returned tus info',
+            'headers': {
+                'Tus-Version': {
+                    'type': 'string'
+                },
+                'Tus-Resumable': {
+                    'type': 'string'
+                },
+                'Tus-Max-Size': {
+                    'type': 'integer'
+                },
+                'Tus-Extension': {
+                    'type': 'string'
+                }
+            }
+        }
+    }))
 class TusOptionsFile(DefaultOPTIONS, UploadFile):
 
     async def render(self):